Restructure folders to better handle styles
