+++ src/components/Toolbar/Toolbar.scss
@@ -1,0 +1,1 @@
+@import '../../utilities/variables.scss';
