+++ src/main.scss
@@ -1,0 +1,18 @@
+@import './utilities/variables.scss';
+@import './components/Epic/Epic.scss';
+@import './components/Lane/Lane.scss';
+@import './components/Toolbar/Toolbar.scss';
+
+.lane {
+    width: 20vw
+}
+
+.laneTitle {
+    width: 250px;
+    height: 80px;
+    border: 1px solid #D6B656;
+    background-color: #FFF2CC;
+    text-align: center;
+    margin: auto
+}
+
