+++ src/utilities/variables.scss
@@ -1,0 +1,2 @@
+$primary-color: #333;
+$inProgress: #FFF2CC;
